Refactoring frontend
* Change to use PositionChangeLog for history information view
* Renaming
